+++ src/compas_fab/ghpython/components/Cf_VisualizeRobot/code.py
@@ -5,6 +5,7 @@ COMPAS FAB v0.18.2
 """
 from compas.geometry import Frame
 from compas.geometry import Transformation
+from compas_ghpython.artists import FrameArtist
 from compas_ghpython.artists import MeshArtist
 from ghpythonlib.componentbase import executingcomponent as component
 
@@ -21,11 +22,11 @@ class RobotVisualize(component):
         ee_frame = None
 
         if robot:
-            show_visual = show_visual or True
+            show_visual = True if show_visual is None else show_visual
             configuration = configuration or robot.zero_configuration()
 
             robot.update(configuration, visual=show_visual, collision=show_collision)
-            frames = robot.transformed_frames(configuration, group)
+            compas_frames = robot.transformed_frames(configuration, group)
 
             if show_visual:
                 visual = robot.artist.draw_visual()
@@ -34,13 +35,21 @@ class RobotVisualize(component):
                 collision = robot.artist.draw_collision()
 
             if show_base_frame:
-                base_frame = frames[0]
+                base_compas_frame = compas_frames[0]
+                artist = FrameArtist(base_compas_frame)
+                base_frame = artist.draw()
 
             if show_end_effector_frame:
-                ee_frame = robot.forward_kinematics(configuration, group, options=dict(solver='model'))
+                ee_compas_frame = robot.forward_kinematics(configuration, group, options=dict(solver='model'))
+                artist = FrameArtist(ee_compas_frame)
+                ee_frame = artist.draw()
 
-            if not show_frames:
-                frames = None
+            if show_frames:
+                frames = []
+                for compas_frame in compas_frames[1:]:
+                    artist = FrameArtist(compas_frame)
+                    frame = artist.draw()
+                    frames.append(frame)
 
             if show_acm:
                 attached_meshes = []